backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart



requests
pytest
httpx
sqlmodel
google-generativeai
python-dotenv
